% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{add_coclust_score}
\alias{add_coclust_score}
\title{Calculate co-clustering scores for values of k}
\usage{
add_coclust_score(bt, ks = c(2, 5, 10),
  parallel = getOption("tglkmeans.parallel"))
}
\arguments{
\item{bt}{output of \code{bootclust}}

\item{ks}{values of k to calculate scores}

\item{parallel}{calcualte scores parallely}
}
\value{
\code{bt} with additional coclust_score field that contains a tibble with score for each cluster in for each value of \code{ks}
}
\description{
Calculate co-clustering scores for values of k
}
\examples{
data <- simulate_data(n = 200, sd = 0.6, nclust = 5, dims = 2, add_true_clust = TRUE)
bt <- bootclust(data \%>\% select(id, starts_with("V")), k_boot = 5, N_boot = 100, boot_ratio = 0.75)
bt <- add_coclust_score(bt, ks = c(2, 3, 5))
plot_coclust_scores(bt)
}
