% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_kmeans}
\alias{bootstrap_kmeans}
\title{Bootstrapping}
\usage{
bootstrap_kmeans(df, k, N_boot, boot_ratio = 0.75,
  parallel = getOption("tglkmeans.parallel"), id_column = TRUE, ...)
}
\arguments{
\item{df}{data frame. Each row is a single observation and each column is a dimension.
the first column can contain id for each observation (if id_column is TRUE).}

\item{k}{number of clusters}

\item{N_boot}{number of bootstrapping iterations}

\item{boot_ratio}{percent of observations to sample on each iteration}

\item{parallel}{run parallely (using doMC backend)}

\item{id_column}{\code{df}'s first column contains the observation id}

\item{...}{additional parameters to TGL_kmeans}
}
\value{
list with the following components:
\describe{
  \item{coclust:}{NxN matrix (where N is the number of observations) with the number of times observation i and j occured in the same cluster.}
  \item{num_trials:}{NxN matrix with the number of times observation i and j where sampled together.}
  \item{coclust_frac:}{fraction of times observation i and j where clustered together out of the times they were sampled together (coclust matrix divided by num_trails matrix).}
}.
}
\description{
Bootstrapping
}
\examples{
d <- simulate_data(nclust=6)
bootstrap <- bootstrap_kmeans(d, k=6, N_boot=100)
names(bootstrap)
bootstrap$coclust[1:5, 1:5]
bootstrap$num_trials[1:5, 1:5]
bootstrap$coclust_frac[1:5, 1:5]

bootstrap_kmeans(d, k=6, N_boot=100, tidy=TRUE)


}
