% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{cutree_bootclust}
\alias{cutree_bootclust}
\title{Cutree of bootstrap clustering}
\usage{
cutree_bootclust(bt, k, min_coclust = 0.5, tidy = FALSE)
}
\arguments{
\item{bt}{output of \code{bootclust}}

\item{k}{an integer scalar or vector with the desired number of groups}

\item{min_coclust}{minimal co-clust score for observation in a cluster}

\item{tidy}{tidy output}
}
\description{
Cutree of bootstrap clustering
}
