% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{match_clusters}
\alias{match_clusters}
\title{Match clusters to true clusters}
\usage{
match_clusters(data, res, nclust)
}
\arguments{
\item{data}{data frame with 'id' and 'true_clust' columns}

\item{res}{a clustering result that has a 'clust' tibble with 'id' and 'clust' columns
(e.g. attach \code{res$clust <- tibble(id = ..., clust = ...)} to a \code{stats::kmeans} result;
for \code{\link{TGL_kmeans_tidy}} output the 'cluster' tibble is matched by partial name)}

\item{nclust}{number of clusters}
}
\value{
data frame with matched clusters
}
\description{
Match clusters to true clusters
}
