% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot_coclust_cutree_mat}
\alias{plot_coclust_cutree_mat}
\title{Plot co-clustering matrix after cutree_bootclust}
\usage{
plot_coclust_cutree_mat(bt,
  col = colorRampPalette(rev(RColorBrewer::brewer.pal(11, "RdBu")))(1000),
  ...)
}
\arguments{
\item{bt}{output of \code{cutree_bootclust}}

\item{col}{color pallete}
}
\value{
None
}
\description{
Plot co-clustering matrix after cutree_bootclust
}
\examples{
data <- simulate_data(n=200, sd=0.6, nclust=5, dims=2, add_true_clust=TRUE)
bt <- bootclust(data \%>\% select(id, starts_with('V')), k=5, N_boot=100, boot_ratio=0.75)
bt <- cutree_bootclust(bt, min_coclust=0.7, k=5)
plot_coclust_cutree_mat(bt)
}
