% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot_coclust_mat}
\alias{plot_coclust_mat}
\title{Plot co-clustering matrix}
\usage{
plot_coclust_mat(bt, col = colorRampPalette(rev(RColorBrewer::brewer.pal(11,
  "RdBu")))(1000), ...)
}
\arguments{
\item{bt}{output of \code{bootclust}}

\item{col}{color pallete}
}
\value{
None
}
\description{
Plot co-clustering matrix
}
