% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot_coclust_score}
\alias{plot_coclust_score}
\title{Plot co-clustering score for different choices of k}
\usage{
plot_coclust_score(coclust_score, ks = c(2, 4, 5, 6, 10, 15), fig_fn = NULL,
  ...)
}
\arguments{
\item{coclust_score}{\code{bt$coclust_score}) where \code{bt} was returned from \code{bootclust}}

\item{ks}{k values to plot}

\item{fig_fn}{filename of the output figure. if NULL figure would be plotted to the screen}

\item{...}{additional paramters to \code{ggplot2::ggsave}}
}
\value{
ggplot object with the densities of k
}
\description{
Plot co-clustering score for different choices of k
}
