% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate normal data for kmeans tests}
\usage{
simulate_data(n = 100, sd = 0.3, nclust = 30, dims = 2,
  frac_na = NULL)
}
\arguments{
\item{n}{number of observations per cluster}

\item{sd}{sd}

\item{nclust}{number of clusters}

\item{dims}{number of dimensions}

\item{frac_na}{fraction of NA in the first dimension}
}
\value{
simulated data
}
\description{
Simulate normal data for kmeans tests
}
\examples{
simluate_data()

}
