% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_clustering}
\alias{test_clustering}
\title{Test clustering performance}
\usage{
test_clustering(n, sd, nclust, dims = 2, method = "euclid", frac_na = NULL)
}
\arguments{
\item{n}{number of observations per cluster}

\item{sd}{standard deviation}

\item{nclust}{number of clusters}

\item{dims}{number of dimensions}

\item{method}{distance metric}

\item{frac_na}{fraction of NA values}
}
\value{
fraction of correctly classified observations
}
\description{
Test clustering performance
}
