% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TGL_kmeans_cpp}
\alias{TGL_kmeans_cpp}
\title{Run kmeans}
\usage{
TGL_kmeans_cpp(ids, mat, k, metric, max_iter = 40, min_delta = 0.0001,
  random_seed = TRUE, seed = -1L)
}
\arguments{
\item{df}{data frame}
}
\description{
Run kmeans
}
