% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tglkmeans-package.R
\docType{package}
\name{tglkmeans-package}
\alias{tglkmeans}
\alias{tglkmeans-package}
\title{tglkmeans: Efficient Implementation of K-Means++ Algorithm}
\description{
Efficient implementation of K-Means++ algorithm. For more information see (1) "kmeans++ the advantages of the k-means++ algorithm" by David Arthur and Sergei Vassilvitskii (2007), Proceedings of the eighteenth annual ACM-SIAM symposium on Discrete algorithms, Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, pp. 1027-1035, and (2) "The Effectiveness of Lloyd-Type Methods for the k-Means Problem" by Rafail Ostrovsky, Yuval Rabani, Leonard J. Schulman and Chaitanya Swamy \doi{10.1145/2395116.2395117}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tanaylab.github.io/tglkmeans/}
  \item \url{https://github.com/tanaylab/tglkmeans}
  \item Report bugs at \url{https://github.com/tanaylab/tglkmeans/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Amos Tanay
}

Other contributors:
\itemize{
  \item Weizmann Institute of Science [copyright holder]
}

}
\keyword{internal}
