% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tglkmeans.set_parallel}
\alias{tglkmeans.set_parallel}
\title{Set parallel threads}
\usage{
tglkmeans.set_parallel(thread_num)
}
\arguments{
\item{thread_num}{number of threads. use '1' for non parallel behavior}
}
\value{
None
}
\description{
Set parallel threads
}
\examples{
\donttest{
tglkmeans.set_parallel(8)
}
}
