% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootclust}
\alias{bootclust}
\title{Bootstrap clustering}
\usage{
bootclust(df, N_boot, boot_ratio = 0.75, k_boot = NULL,
  bootstrap_func = "bootstrap_kmeans", max_k = NULL,
  heatmap_plot_fn = NULL, width = 700, height = 700, device = "png",
  ...)
}
\arguments{
\item{df}{data frame. Each row is a single observation and each column is a dimension. Input to \code{bootstrap_func} (first paramter).}

\item{N_boot}{number of bootstraps.}

\item{boot_ratio}{fraction of observations to sample in each bootstrap.}

\item{k_boot}{k to use in \code{bootstrap_func}(\code{k} parameter).}

\item{bootstrap_func}{function to use bootstrapping. Should take \code{df} as the first parameter, have the following parameters: \code{k}, \code{boot_ratio}, \code{N_boot}, and return a list with \code{coclust}, \code{num_trials} and \code{coclust_frac}.}

\item{max_k}{maximal k to test. if NULL, would be chosen as \code{floor(nrow(df) / 40)}}

\item{heatmap_plot_fn}{filename for coclust heatmap.}

\item{width}{width for heatmap plot.}

\item{height}{height for heatmap plot.}

\item{device}{device for heatmap plot (default \code{png}),}

\item{...}{other paramters to \code{bootstrap_func}.}
}
\value{
list with the following components:
\describe{
  \item{mat:}{original \code{df}. NxM}
  \item{coclust:}{xN matrix with number of times item \code{i} was sampled with item \code{j} in the same cluster.}
  \item{num_trials:}{NxN matrix with number of times item \code{i} was sampled with item \code{j}.}
  \item{coclust_frac:}{NxN matrix with \code{coclust} / \code{num_trials}.}
  \item{cm:}{NxN correlation matrix.}
  \item{hc:}{hierachical clustering of \code{coclust_frac} matrix.}
  \item{coclust_score:}{tibble with score for each cluster in each k from 1 to \code{max_k}.}
}
}
\description{
Bootstrap clustering
}
