% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{cutree_bootclust}
\alias{cutree_bootclust}
\title{Cutree of bootstrap clustering}
\usage{
cutree_bootclust(bt, k, min_coclust = 0.5, tidy = FALSE)
}
\arguments{
\item{bt}{output of \code{bootclust}}

\item{k}{an integer scalar or vector with the desired number of groups}

\item{min_coclust}{minimal co-clust score for observation in a cluster}

\item{tidy}{tidy output}
}
\value{
if \code{tidy == TRUE}: \code{bt} with the following aditional fields:
\describe{
  \item{clust:}{tibble with `id` column with the observation id (`1:n` if no id column was supplied), and `clust` column with the observation assigned cluster.}
  \item{centers:}{tibble with `clust` column and the cluster centers.}
  \item{size:}{tibble with `clust` column and `n` column with the number of points in each cluster.}
  \item{score:}{}
}
if \code{tidy == FALSE}: \code{bt} with the following aditional fields:
\describe{
  \item{cluster:}{A vector of integers (from ‘1:k’) indicating the cluster to which each point is allocated.}
  \item{centers:}{A matrix of cluster centres.}
  \item{size:}{The number of points in each cluster.}
  \item{score:}{}
}
}
\description{
Cutree of bootstrap clustering
}
