% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample_matrix}
\alias{downsample_matrix}
\title{Downsample a matrix to a target number of in each column}
\usage{
downsample_matrix(mat, target_n, seed = NULL, remove_columns = FALSE)
}
\arguments{
\item{mat}{The input matrix to be downsampled. Can be a matrix or sparse matrix (dgCMatrix). If the matrix contains NAs, the function will run significantly slower.}

\item{target_n}{The target number of samples to downsample to}

\item{seed}{The random seed for reproducibility (default is NULL)}

\item{remove_columns}{Logical indicating whether to remove columns with small sums (default is FALSE)}
}
\value{
The downsampled matrix
}
\description{
This function takes a matrix and downsamples it to a target number of samples.
It uses a random seed for reproducibility and allows for removing columns with
small sums.
}
\examples{
mat <- matrix(1:12, nrow = 4)
downsample_matrix(mat, 2)

# Remove columns with small sums
downsample_matrix(mat, 12, remove_columns = TRUE)

# sparse matrix
mat_sparse <- Matrix::Matrix(mat, sparse = TRUE)
downsample_matrix(mat_sparse, 2)

}
