% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{match_clusters}
\alias{match_clusters}
\title{Match clusters to true clusters}
\usage{
match_clusters(data, res, nclust)
}
\arguments{
\item{data}{data frame with 'id' and 'true_clust' columns}

\item{res}{result from TGL_kmeans_tidy (must have a 'cluster' tibble with 'id' and 'clust' columns)}

\item{nclust}{number of clusters}
}
\value{
data frame with matched clusters
}
\description{
Match clusters to true clusters
}
