% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot_coclust_mat}
\alias{plot_coclust_mat}
\title{Plot co-clustering matrix}
\usage{
plot_coclust_mat(bt, col = colorRampPalette(rev(RColorBrewer::brewer.pal(11,
  "RdBu")))(1000), ...)
}
\arguments{
\item{bt}{output of \code{bootclust}}

\item{col}{color pallete}
}
\value{
None
}
\description{
Plots a heatmap of coclust_frac: \code{coclust / num_trials}
}
\examples{
data <- simulate_data(n=200, sd=0.6, nclust=5, dims=2, add_true_clust=TRUE)
bt <- bootclust(data \%>\% select(id, starts_with('V')), k_boot=5, N_boot=100, boot_ratio=0.75)
plot_coclust_mat(bt)
}
