% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot_coclust_score}
\alias{plot_coclust_score}
\title{Plot co-clustering score for different choices of k}
\usage{
plot_coclust_score(bt, ks = NULL, fig_fn = NULL, ...)
}
\arguments{
\item{bt}{output of \code{bootclust}}

\item{ks}{k values to plot}

\item{fig_fn}{filename of the output figure. if NULL figure would be plotted to the screen}

\item{...}{additional paramters to \code{ggplot2::ggsave}}
}
\value{
ggplot object with the densities of k
}
\description{
Plot co-clustering score for different choices of k
}
\examples{
data <- simulate_data(n=200, sd=0.6, nclust=5, dims=2, add_true_clust=TRUE)
bt <- bootclust(data \%>\% select(id, starts_with('V')), k=5, N_boot=100, boot_ratio=0.75)
plot_coclust_score(bt, ks=2:5)
}
